requirements.txt
luma.led-matrix~=1.5.0
loguru~=0.6.0
RPi.GPIO~=0.7.1